APIs
userController APIs started
divorceController APIs started

+++ src/main/java/gr/hua/dit/ds/divorce/it22047_it22113_it22047/repositories/DivorceRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface DivorceRepository extends JpaRepository<Divorce, Long> {
+public interface DivorceRepository extends JpaRepository<Divorce, Integer> {
